the largest prime factor of the number

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
   */
 int main(void)
 {
-	long x, factor;
+	int x, factor;
 	long n = 612852475143;
 	double square = sqrt(n);
 
@@ -16,6 +16,6 @@ int main(void)
 		if (n % x == 0)
 			factor = n / x;
 	}
-	printf("%ld\n", factor);
+	printf("%d\n", factor);
 	return (0);
 }
